requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==2.1.7
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-rest-knox==3.6.0
djangorestframework==3.9.2
gunicorn==19.9.0
Pillow==5.4.1
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
